// demos/expense-approval-automation/src/generated/models/Expaa_approvalsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Expaa_approvalsexpaa_action = {
  1: 'Submitted',
  2: 'Approved',
  3: 'Rejected',
  4: 'Returned',
  5: 'Reimbursed'
} as const;
export type Expaa_approvalsexpaa_action = keyof typeof Expaa_approvalsexpaa_action;
export const Expaa_approvalsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Expaa_approvalsstatecode = keyof typeof Expaa_approvalsstatecode;
export const Expaa_approvalsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Expaa_approvalsstatuscode = keyof typeof Expaa_approvalsstatuscode;

export interface Expaa_approvalsBase {
  expaa_action?: Expaa_approvalsexpaa_action;
  expaa_actionon?: string;
  expaa_approvalid: string;
  expaa_approvername?: string;
  "expaa_expenseid@odata.bind"?: string;
  expaa_name?: string;
  expaa_notes?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Expaa_approvalsstatecode;
  statuscode?: Expaa_approvalsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Expaa_approvals extends Expaa_approvalsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  expaa_actionname?: string;
  expaa_expenseidname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  expaa_expenseid?: object;
  _expaa_expenseid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}